Add script and result for testing the script std_avrg_using_cdo.py.
This commit is done for testing the global annual averaging script std_avrg_using_cdo.py against the general_proc.sh script from S. Ferrachat.

The file out_diffcdo.txt show the difference in the annual mean time series. The difference is in the prder of 1e-7, occuring most often on Fnet, Fnet_surf and Fnet_toa_min_surf . Theses variables are computed from standard output variables and eh averaging is not done at the same time in the process, this is why the differences occur.